fix context b rarely complete

+++ asymmetric/gameFilter.py
@@ -32,6 +32,8 @@ class AbstractSystem():
 class FilterGameResult(aenum.Enum):
     _settings_ = aenum.NoAlias
     PASS=0.0
+    NO_CODE_FOUND=-1.0
+    CANT_EXTRACT_CLASS=-1.0
     ERROR=-1.0
     TIMEOUT=-1.0
     MOVES_NOT_STRINGIFIABLE=-1.0
@@ -97,6 +99,31 @@ def _filterGame_worker(GameClass, q: Queue):
     ret = _filterGame(GameClass)
     q.put(ret.name)
 
+import re
+def extract_game(text: str) -> Optional[str]:
+    code_blocks= re.findall(r"```(?:python)?\n(.*?)\n```", text, re.DOTALL | re.IGNORECASE)
+    if code_blocks:
+        return code_blocks[-1].strip()
+    return text.split('</think>')[-1].strip()
+
+def run_game_code_and_get_class(game_src: str):
+    """
+    Exec the generated Game class in a minimal namespace that already defines AbstractSystem.
+    Returns Game class object. If not found or doesn't pass check, return None
+    """
+    namespace = {'AbstractSystem': AbstractSystem}
+    try:
+        exec(game_src, namespace, namespace)
+    except Exception as e:
+        print(f'Error executing game code: {e}')
+        return None
+    Game = namespace.get("System", None)
+    if Game is None:
+        print("Generated code did not define class System.")
+        return None
+    return Game
+
+
 LOG_PATH="logs/gameFilter.log"
 import os, inspect, traceback
 
@@ -106,29 +133,40 @@ os.makedirs(os.path.dirname(LOG_PATH), exist_ok=True)
 resultCounter={}
 count=0
 
-def filterGame(GameClass: AbstractSystem, gameCode:str)->FilterGameResult:
+def filterGame(response:str)->FilterGameResult:
     global count
     count+=1
     ret=None
-    try:
-        q = Queue()
-        p = Process(target=_filterGame_worker, args=(GameClass, q))
-        p.start()
-        p.join(5)
-        if p.is_alive():
-            p.terminate()
-            p.join()
-            ret=FilterGameResult.TIMEOUT
-        else:
-            name = q.get_nowait()
-            ret = FilterGameResult[name]
-    except Exception as e:
-        stack=traceback.format_exc()
-        print(f"Error occurred while filtering game {GameClass}:\n{stack}")
-        with open(LOG_PATH,"a") as f:
-            f.write(f"ID:{count}, error stack:\n{stack}\n")
-        ret=FilterGameResult.ERROR
-    code=gameCode
+    gameCode=extract_game(response)
+    if gameCode is None:
+        ret=FilterGameResult.NO_CODE_FOUND
+    else:
+        GameClass=run_game_code_and_get_class(gameCode)
+        if GameClass is None:
+            ret=FilterGameResult.CANT_EXTRACT_CLASS
+    if ret is None:
+        try:
+            q = Queue()
+            p = Process(target=_filterGame_worker, args=(GameClass, q))
+            p.start()
+            p.join(5)
+            if p.is_alive():
+                p.terminate()
+                p.join()
+                ret=FilterGameResult.TIMEOUT
+            else:
+                name = q.get_nowait()
+                ret = FilterGameResult[name]
+        except Exception as e:
+            stack=traceback.format_exc()
+            print(f"Error occurred while filtering game {GameClass}:\n{stack}")
+            with open(LOG_PATH,"a") as f:
+                f.write(f"ID:{count}, error stack:\n{stack}\n")
+            ret=FilterGameResult.ERROR
+    code=gameCode if gameCode is not None else "<no code>"
+    if len(code)>10000: # probably repeating something so not useful to log all
+        code=code.replace('\n','\\n') # reduce lines
+        code=code[:500]+"\n...\n"+code[-500:]
     resultCounter[ret.name]=resultCounter.get(ret.name,0)+1
     with open(LOG_PATH,"a") as f:
         f.write(f"ID:{count}, Code:\n```\n{code}\n```\nResult: {ret.name}\n")
